refactor(vscode): break registry<->register import cycle; no-cycle=error
Extract tool metadata + isToolEnabled/isEnabled into a new leaf module
extension/toolEnabled.ts (TOOL_META). toolRegistry.ts keeps only the
register-function wiring (REGISTER_BY_ID, keyed by id) + the activate/watch
lifecycle, composing full descriptors from TOOL_META. The 5 tools/*/register.ts
now import isToolEnabled from the leaf instead of back into toolRegistry, so the
toolRegistry -> register -> toolRegistry cycle is gone (oxlint import/no-cycle:
10 -> 0, vscode typecheck green). Live enable/disable behavior unchanged
(settingKey/contextKey mirroring, liveToggle) — same logic, relocated metadata.

With the last cycle gone, escalate .oxlintrc.json import/no-cycle warn -> error.
